Step by step homepage URL: Relative URL instead of Absolute URL.
The Step by step page carries a URL pointing back to its parent which is a Step
by step homepage.  This was an absolute URL.  Now it should be a relative URL.

A relative URL is necessary because the URL of our reverse proxy is different
from the Drupal site.