Work around for conditional display of theme region.
Rendering the sidebars in the page preprocess function is breaking blocks.  For
example, any asset library or drupalSettings **attached** to a block's render
array is disappearing.  To avoid this, we should render a **copy** of the
sidebars' render array in the page preprocess function to determine if the
sidebars will be empty or not.