Bug fix for page template suggestion.
For node pages, the current suggestion is adding a content type specific
template suggestion as the final one.  This overrides Drupal's front-page
specific template suggestion making it impossible to have a frontpage
specific page template.

This change addresses this by restoring the frontpage specific template
suggestion to the final position.